(PC-29690)[PRO] refactor: extract url params parser to use it in analytics logger
